Final fixes and restructuring Query modules to fix standard Ruby conventions
