+++ lib/neo4j-core.rb
@@ -9,8 +9,7 @@ require 'neo4j/property_validator'
 require 'neo4j/property_container'
 require 'neo4j-core/active_entity'
 require 'neo4j-core/helpers'
-require 'neo4j-core/query_find_in_batches'
-require 'neo4j-core/query'
+require 'neo4j/core/query'
 
 require 'neo4j/entity_equality'
 require 'neo4j/node'
