+++ lib/neo4j/core/query.rb
@@ -1,4 +1,6 @@
-require 'neo4j-core/query_clauses'
+require 'neo4j/core/query/clauses'
+require 'neo4j/core/query/find_in_batches'
+require 'neo4j/core/query/parameters'
 require 'active_support/notifications'
 
 module Neo4j
@@ -12,54 +14,11 @@ module Neo4j
     # See also the following link for full cypher language documentation:
     # http://docs.neo4j.org/chunked/milestone/cypher-query-lang.html
     class Query
-      include Neo4j::Core::QueryClauses
-      include Neo4j::Core::QueryFindInBatches
+      include Neo4j::Core::Query::Clauses
+      include Neo4j::Core::Query::FindInBatches
       DEFINED_CLAUSES = {}
 
       attr_accessor :clauses
-
-      class Parameters
-        def initialize(hash = nil)
-          @parameters = (hash || {})
-        end
-
-        def to_hash
-          @parameters
-        end
-
-        def copy
-          self.class.new(@parameters.dup)
-        end
-
-        def add_param(key, value)
-          free_param_key(key).tap do |k|
-            @parameters[k.freeze] = value
-          end
-        end
-
-        def remove_param(key)
-          @parameters.delete(key.to_sym)
-        end
-
-        def add_params(params)
-          params.map do |key, value|
-            add_param(key, value)
-          end
-        end
-
-        private
-
-        def free_param_key(key)
-          k = key.to_sym
-
-          return k if !@parameters.key?(k)
-
-          i = 2
-          i += 1 while @parameters.key?("#{key}#{i}".to_sym)
-
-          "#{key}#{i}".to_sym
-        end
-      end
 
       class << self
         attr_accessor :pretty_cypher
@@ -195,7 +154,7 @@ module Neo4j
       # @example
       #    # Creates a query representing the cypher: MATCH (n:Person) SET n.age = 19
       #    Query.new.match(n: :Person).set_props(n: {age: 19})
-      def set_props(*args)
+      def set_props(*args) # rubocop:disable Style/AccessorMethodName
         build_deeper_query(SetClause, args, set_props: true)
       end
 
@@ -335,6 +294,8 @@ module Neo4j
       EMPTY = ' '
       NEWLINE = "\n"
       def to_cypher(options = {})
+        separator = options[:pretty] ? NEWLINE : EMPTY
+
         cypher_string = partitioned_clauses.map do |clauses|
           clauses_by_class = clauses.group_by(&:class)
 
@@ -343,8 +304,8 @@ module Neo4j
           end
 
           cypher_parts.compact!
-          cypher_parts.join(options[:pretty] ? NEWLINE : EMPTY).tap(&:strip!)
-        end.join(options[:pretty] ? NEWLINE : EMPTY)
+          cypher_parts.join(separator).tap(&:strip!)
+        end.join(separator)
 
         cypher_string = "CYPHER #{@options[:parser]} #{cypher_string}" if @options[:parser]
         cypher_string.tap(&:strip!)
@@ -392,9 +353,7 @@ module Neo4j
 
       def clause?(method)
         clause_class = DEFINED_CLAUSES[method] || CLAUSIFY_CLAUSE.call(method)
-        clauses.any? do |clause|
-          clause.is_a?(clause_class)
-        end
+        clauses.any? { |clause| clause.is_a?(clause_class) }
       end
 
       protected
@@ -406,9 +365,7 @@ module Neo4j
       end
 
       def remove_clause_class(clause_class)
-        @clauses = @clauses.reject do |clause|
-          clause.is_a?(clause_class)
-        end
+        @clauses = @clauses.reject { |clause| clause.is_a?(clause_class) }
       end
 
       private
@@ -444,7 +401,7 @@ module Neo4j
               second_to_last << clause
             elsif clause_is_with_following_order_or_limit?(clause)
               second_to_last << clause
-              second_to_last.sort_by! { |c| c.is_a?(::Neo4j::Core::QueryClauses::OrderClause) ? 1 : 0 }
+              second_to_last.sort_by! { |c| c.is_a?(::Neo4j::Core::Query::Clauses::OrderClause) ? 1 : 0 }
             else
               @partitioning.last << clause
             end
@@ -464,18 +421,18 @@ module Neo4j
         def clause_is_order_or_limit_directly_following_with_or_order?(clause)
           self.class.clause_is_order_or_limit?(clause) &&
             @partitioning[-2] &&
-            (@partitioning[-2].last.is_a?(::Neo4j::Core::QueryClauses::WithClause) ||
-              @partitioning[-2].last.is_a?(::Neo4j::Core::QueryClauses::OrderClause))
+            (@partitioning[-2].last.is_a?(::Neo4j::Core::Query::Clauses::WithClause) ||
+              @partitioning[-2].last.is_a?(::Neo4j::Core::Query::Clauses::OrderClause))
         end
 
         def clause_is_with_following_order_or_limit?(clause)
-          clause.is_a?(::Neo4j::Core::QueryClauses::WithClause) &&
+          clause.is_a?(::Neo4j::Core::Query::Clauses::WithClause) &&
             @partitioning[-2] && @partitioning[-2].any? { |c| self.class.clause_is_order_or_limit?(c) }
         end
 
         def self.clause_is_order_or_limit?(clause)
-          clause.is_a?(::Neo4j::Core::QueryClauses::OrderClause) ||
-            clause.is_a?(::Neo4j::Core::QueryClauses::LimitClause)
+          clause.is_a?(::Neo4j::Core::Query::Clauses::OrderClause) ||
+            clause.is_a?(::Neo4j::Core::Query::Clauses::LimitClause)
         end
       end
