+++ lib/neo4j/core/query/find_in_batches.rb
@@ -1,0 +1,51 @@
+module Neo4j
+  module Core
+    class Query
+      module FindInBatches
+        def find_in_batches(node_var, prop_var, options = {})
+          validate_find_in_batches_options!(options)
+
+          batch_size = options.delete(:batch_size) || 1000
+
+          query = reorder(node_var => prop_var).limit(batch_size)
+
+          records = query.to_a
+
+          while records.any?
+            records_size = records.size
+            primary_key_offset = primary_key_offset(records.last, node_var, prop_var)
+
+            yield records
+
+            break if records_size < batch_size
+
+            records = query.where("#{node_var}.#{prop_var} > {primary_key_offset}").params(primary_key_offset: primary_key_offset).to_a
+          end
+        end
+
+        def find_each(*args)
+          find_in_batches(*args) do |batch|
+            batch.each { |result| yield result }
+          end
+        end
+
+        private
+
+        def validate_find_in_batches_options!(options)
+          invalid_keys = options.keys.map(&:to_sym) - [:batch_size]
+          fail ArgumentError, "Invalid keys: #{invalid_keys.join(', ')}" if not invalid_keys.empty?
+        end
+
+        def primary_key_offset(last_record, node_var, prop_var)
+          last_record.send(node_var).send(prop_var)
+        rescue NoMethodError
+          begin
+            last_record.send(node_var)[prop_var.to_sym]
+          rescue NoMethodError
+            last_record.send("#{node_var}.#{prop_var}") # In case we're explicitly returning it
+          end
+        end
+      end
+    end
+  end
+end
